requirements.txt
folium
holidays
imbalanced-learn
joblib
lightgbm
matplotlib
numpy
pandas
plotly
pyarrow
pytest
requests
scikit-learn
seaborn
streamlit
streamlit-extras
streamlit-folium
xgboost
mysql-connector-python
python-dotenv